PostgreSQL Example Query
The use of PostgreSQL is demonstrated in this portfolio piece.  Variations between Standard and the PostgreSQL dialect are illustratively identified -- including the use of single-quotation and double-quotation within queries.  Tasks and objectives are labeled and included as "--" commentary above the query coding.  PLEASE NOTE: "--" are included in prior SELECT queries to offset execution when progressing through the tasks during this particular project.